+++ docs/controllers/user_controller/sidebar-items.js
@@ -1,0 +1,1 @@
+window.SIDEBAR_ITEMS = {"fn":["get_all_users_controller","insert_single_user_controller"]};