+++ docs/models/task/sidebar-items.js
@@ -1,0 +1,1 @@
+window.SIDEBAR_ITEMS = {"struct":["NewTask","PatchTask","PutTask","Task"]};