+++ docs/schema/users/columns/sidebar-items.js
@@ -1,0 +1,1 @@
+window.SIDEBAR_ITEMS = {"struct":["avatar_url","bio","created_at","email","fullname","mobile_phone","password","role","star","updated_at","user_id","username"]};