+++ CHANGELOG.md
@@ -5,9 +5,9 @@
     - Initiate project, and finish project structure building.
 - 2024.04.05:
   - `16e7b5e52d3808252a392be375d9df6ee77a1fc7`
-    - Finish GET, POST, PATCH, DELETE.
+    - Finish `GET`, `POST`, `PATCH`, `DELETE`.
 - 2024.04.06:
   - `daaa418b093be85ae2c99f03b8d4e038cabfc575`
-    - Finish Task CRUD.
+    - Finish `Task` CRUD.
   - `7c76dac6b5e34ae2d710555ed1fc5a40195b82ac`
-    - Add `users, follows, posts` table, finish User structure and implemented some GET and POST for User.
+    - Add `users, follows, posts` table, finish `User` structure and implemented some `GET` and `POST` for `User`.